requirements.txt
python-whois==0.8.0
dnspython==2.4.2
requests==2.31.0